requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
marshmallow
marshmallow-sqlalchemy
psycopg2-binary
python-dotenv
Werkzeug
python-dateutil